# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Time Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/task-assignments:
    get:
      operationId: listProjectTaskAssignments
      tags: [tasks]
      summary: Who is doing what on a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectTaskAssignment"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: assignProjectTask
      tags: [tasks]
      summary: Assign a task to somebody on the project (Associate and above)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectTaskAssignmentInput"
      responses:
        "201":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTaskAssignmentRecord"
        "400":
          description: Task not enabled here, or the person is not on the project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: An analyst tried to assign work
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already assigned to that person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /task-assignments/{assignmentId}:
    delete:
      operationId: deleteProjectTaskAssignment
      tags: [tasks]
      summary: Withdraw an assignment (Associate and above)
      parameters:
        - $ref: "#/components/parameters/AssignmentId"
      responses:
        "204":
          description: Withdrawn
        "403":
          description: An analyst tried to withdraw an assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProjectTaskAssignment:
      type: object
      required: [id, taskId, taskName, assigneeUserId, assigneeName, assigneeRole, assignedById, assignedAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        taskName:
          type: string
        assigneeUserId:
          type: integer
        assigneeName:
          type: string
        assigneeRole:
          type: string
          enum: [analyst, associate, avp, md]
        assigneeTitle:
          type: ["string", "null"]
          description: >-
            Overrides the rank label where the person's designation differs from it. Null shows the ordinary role label.
        assignedById:
          type: integer
        assignedAt:
          type: string
          format: date-time
    ProjectTaskAssignmentRecord:
      type: object
      required: [id, projectId, taskId, assigneeUserId, assignedById, assignedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        taskId:
          type: integer
        assigneeUserId:
          type: integer
        assignedById:
          type: integer
        assignedAt:
          type: string
          format: date-time
    ProjectTaskAssignmentInput:
      type: object
      required: [taskId]
      properties:
        taskId:
          type: integer
        assigneeUserId:
          type: integer
          description: "Omit to take the task yourself."
...