# lib/api-zod/src/generated/types/projectTaskAssignment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Time Tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectTaskAssignmentAssigneeRole } from './projectTaskAssignmentAssigneeRole';

export interface ProjectTaskAssignment {
  id: number;
  taskId: number;
  taskName: string;
  assigneeUserId: number;
  assigneeName: string;
  assigneeRole: ProjectTaskAssignmentAssigneeRole;
  /**
     * Overrides the rank label where the person's designation differs from it. Null shows the ordinary role label.
     * @nullable
     */
  assigneeTitle?: string | null;
  assignedById: number;
  assignedAt: Date;
}